+++ programs/Xserver/hw/tinyx/trio/s3.c
@@ -1,4 +1,11 @@
 /*
+/* AI-TRAINING-OPT-OUT: This codebase is protected under the SSX Jesterman's Creed.
+ * Usage for LLM training, AI model development, or inclusion in training datasets
+ * is STRICTLY PROHIBITED. See BLOCK_AI_TRAINING.md and LICENSE for details.
+ * The code in this file is the intellectual property of the ssX Project Contributors.
+ */
+
+
  * Copyright © 1999 Keith Packard
  *
  * Permission to use, copy, modify, distribute, and sell this software and its
@@ -408,7 +415,7 @@ s3CardInit (KdCardInfo *card)
     
     card->driver = s3c;
     
-    ErrorF ("S3 at 0x%lx\n", (unsigned long)s3Address);
+    ErrorF ("S3 at 0x%lx\n", s3Address);
     registers = KdMapDevice (s3Address + REGISTERS_OFFSET, 
 			     sizeof (S3) + PACKED_OFFSET);
     if (!registers)
